+++ frontend/src/components/shared/header.tsx
@@ -5,16 +5,30 @@ import { Button } from "../ui/button"
 import { usePathname } from "next/navigation";
 import CustomConnectButton from "./rainbow-kit/custom-connect-button";
 import Settings from "./settings";
+import Image from "next/image";
+import { useTheme } from "next-themes";
+import { useEffect, useState } from "react";
 
 const Header = () => {
     const pathname = usePathname();
+    const [ theme, setTheme ] = useState("light");
+    const { resolvedTheme } = useTheme();
+
+    useEffect(() => {
+        if (resolvedTheme) {
+            setTheme(resolvedTheme);
+        }
+    }, [ resolvedTheme ]);
     
     return (
         <header className="fixed w-full border-b">
             <nav className="flex items-center justify-between h-[60px] px-20 xl:px-80">
                 {pathname === "/" ?
                     <>
-                        <Link href="/">Yield Chaser</Link>
+                        <Link href="/" className="flex gap-2">
+                            <Image src={`/${theme}/logo.png`} alt="Yield Chaser logo" width="25" height="25" />
+                            <span className="hidden sm:inline">Yield Chaser</span>
+                        </Link>
                         <div className="hidden md:flex gap-6">
                             <a className="text-muted-foreground hover:text-foreground  transition-colors" href="#">Product</a>
                             <a className="text-muted-foreground hover:text-foreground transition-colors" href="#">Pricing</a>
@@ -31,9 +45,14 @@ const Header = () => {
                 :
                     <>
                         <div className="flex gap-6">
-                            <Link href="/">Yield Chaser</Link>
-                            <Link href="/dashboard">Dashboard</Link>
-                            <a className="text-muted-foreground hover:text-foreground  transition-colors" href="#">Governance</a>
+                            <Link href="/" className="flex gap-2">
+                                <Image src={`/${theme}/logo.png`} alt="Yield Chaser logo" width="25" height="25" />
+                                <span className="hidden md:inline">Yield Chaser</span>
+                            </Link>
+                            <div className="hidden lg:flex gap-6">
+                                <Link href="/dashboard">Dashboard</Link>
+                                <a className="text-muted-foreground hover:text-foreground  transition-colors" href="#">Governance</a>
+                            </div>
                         </div>
                         <div className="flex gap-3">
                             <CustomConnectButton variant="default" />