+++ frontend/src/types/account-type.ts
@@ -4,7 +4,6 @@ export interface AccountType {
     earnedUsdc: number,
     totalUsdc: number,
     eth: number,
-    noReallocationPeriod: number,
     currentVault: `0x${string}`|null
     strategy: `0x${string}`|null
 }